@@ -8,6 +8,7 @@
 import { Component, Input, OnInit, OnDestroy } from '@angular/core';
 import { LiveScoreService } from '../live-score.service';
 import { ActivatedRoute } from '@angular/router';
+import { catchError } from 'rxjs/operators';
 
 @Component({
   selector: 'app-cricket-commentary',
@@ -52,40 +53,44 @@ export class CricketCommentaryComponent implements OnInit, OnDestroy {
 
     initData() {
         if (this.matchId) {
-            // let convMatchId = this.matchId.replace(/:/g, '/');
-            // let url = 'http://synd.cricbuzz.com/j2me/1.0/' + convMatchId;
+            // const convMatchId = this.matchId.replace(/:/g, '/');
+            // const url = 'http://synd.cricbuzz.com/j2me/1.0/' + convMatchId;
             this._liveScoreService.fetchMatchCommentary(this.matchId)
-            .then(res => {
-              this.selectedCommentary = res.json().mchDetails.match[0];
-            })
-            .catch(err => console.error(err));
+                .pipe(
+                    catchError((err: any) => {
+                        console.error(err);
+                        return err;
+                    })
+                ).subscribe((res: any) => {
+                    this.selectedCommentary = res.json().mchDetails.match[0];
+                })
         }
     }
 
     convertGMTtoLocalTime(gmtTime: string) {
-        let time = gmtTime.split(':');
-        if (time.length == 2) {
-          let gmtTimeMin = Math.floor(parseInt(time[0]) * 60) + Math.floor(parseInt(time[1]));
-          let offset = new Date().getTimezoneOffset();
-          let localTimeMin = gmtTimeMin - offset;
+        const time = gmtTime.split(':');
+        if (time.length === 2) {
+          const gmtTimeMin = Math.floor(parseInt(time[0], 10) * 60) + Math.floor(parseInt(time[1], 10));
+          const offset = new Date().getTimezoneOffset();
+          const localTimeMin = gmtTimeMin - offset;
 
           if (localTimeMin < 0 || localTimeMin > (60 * 24)) {
             return gmtTime + ' Hrs GMT';
           }
 
-          let localHours = Math.floor(localTimeMin/60);
-          let localMins = Math.floor(localTimeMin%60);
+          let localHours = Math.floor(localTimeMin / 60);
+          const localMins = Math.floor(localTimeMin % 60);
           let timeDay = 'AM';
           if (localHours >= 12) {
             localHours = localHours - 12;
             timeDay = 'PM';
           }
-          if (localHours == 0) {
+          if (localHours === 0) {
             localHours = 12;
           }
 
-          let hours = (localHours < 10)? ('0' + localHours) : localHours;
-          let mins = (localMins < 10)? ('0' + localMins) : localMins;
+          const hours = (localHours < 10) ? ('0' + localHours) : localHours;
+          const mins = (localMins < 10) ? ('0' + localMins) : localMins;
 
           return ( hours + ':' + mins + ' ' + timeDay);
         }
