@@ -5,58 +5,61 @@
  */
 
 
-import { Injectable } from "@angular/core";
-import { Http, ResponseContentType, Response, RequestOptionsArgs } from "@angular/http";
-import { ServerConfigService } from "../services/server-config.service";
+import { Injectable } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
+
+import { Observable } from 'rxjs';
+
+import { ServerConfigService } from '../services/server-config.service';
 
 
 @Injectable()
 export class LiveScoreService {
 
-    constructor(private _http: Http, private _serverConfigService: ServerConfigService) {
+    constructor(private _http: HttpClient, private _serverConfigService: ServerConfigService) {
     }
 
     //
     // OPERATIONS
     //
 
-    public fetchLiveMatches(): Promise<Response> {
+    public fetchLiveMatches(): Observable<any> {
         return this._http.get(this._serverConfigService.serverUrl
-            + '/api/cricket/livecricscore/csLiveMatches/').toPromise();
+            + '/api/cricket/livecricscore/csLiveMatches/');
     }
 
-    public fetchLiveScore(matchId: string): Promise<Response> {
+    public fetchLiveScore(matchId: string): Observable<any> {
         return this._http.get(this._serverConfigService.serverUrl
-            + '/api/cricket/livecricscore/csLiveScore/?matchId=' + matchId).toPromise();
+            + '/api/cricket/livecricscore/csLiveScore/?matchId=' + matchId);
     }
 
-    public fetchCurrentMatches(): Promise<Response> {
+    public fetchCurrentMatches(): Observable<any> {
         return this._http.get(this._serverConfigService.serverUrl
-            + '/api/cricket/cbLiveMatches/').toPromise();
+            + '/api/cricket/cbLiveMatches/');
     }
 
     // fetch the scorecard
-    public fetchMatchScorecard(matchUrl: string): Promise<Response> {
+    public fetchMatchScorecard(matchUrl: string): Observable<any> {
         return this._http.get(this._serverConfigService.serverUrl
-            + '/api/cricket/cbScorecard/?matchUrl=' + matchUrl).toPromise();
+            + '/api/cricket/cbScorecard/?matchUrl=' + matchUrl);
     }
 
     // fetch the commentary
-    public fetchMatchCommentary(matchUrl: string): Promise<Response> {
+    public fetchMatchCommentary(matchUrl: string): Observable<any> {
         return this._http.get(this._serverConfigService.serverUrl
-            + '/api/cricket/cbCommentary/?matchUrl=' + matchUrl).toPromise();
+            + '/api/cricket/cbCommentary/?matchUrl=' + matchUrl);
     }
 
     // fetch the news headlines
-    public fetchCricketHeadlines(): Promise<Response> {
+    public fetchCricketHeadlines(): Observable<any> {
         return this._http.get(this._serverConfigService.serverUrl
-            + '/api/cricket/cricnews/top-headlines/').toPromise();
+            + '/api/cricket/cricnews/top-headlines/');
     }
 
      // fetch the news headlines
-    public fetchCricketNews(): Promise<Response> {
+    public fetchCricketNews(): Observable<any> {
         return this._http.get(this._serverConfigService.serverUrl
-            + '/api/cricket/cricnews/everything/').toPromise();
+            + '/api/cricket/cricnews/everything/');
     }
 
 }
