@@ -5,7 +5,7 @@
  */
 
 
-import { Injectable } from "@angular/core";
+import { Injectable } from '@angular/core';
 
 
 @Injectable()
